Add All Endpoints, Localization, Mail Service

+++ Vezeeta.Service/Users/DoctorService.cs
@@ -1,9 +1,12 @@
-﻿using System.Linq.Expressions;
+﻿using System.Collections.Generic;
+using System.Linq.Expressions;
 using Vezeeta.Core.Consts;
 using Vezeeta.Core.Domain.Bookings;
+using Vezeeta.Core.Domain.Lookup;
 using Vezeeta.Core.Domain.Users;
 using Vezeeta.Core.Enums;
 using Vezeeta.Core.Repository;
+using Vezeeta.Core.Service;
 using Vezeeta.Core.Service.Users;
 using Vezeeta.Core.Shared;
 
@@ -14,15 +17,21 @@ namespace Vezeeta.Service.Users
         private readonly IDoctorRepository _doctorRepository;
         private readonly IBookingRepository _bookingRepository;
         private readonly IUserService _userService;
+        private readonly IMailService _mailService;
 
-        public DoctorService(IDoctorRepository doctorRepository, IUserService userService, IBookingRepository bookingRepository)
+        public DoctorService(
+            IDoctorRepository doctorRepository, 
+            IUserService userService,
+            IMailService mailService,
+            IBookingRepository bookingRepository)
         {
             _doctorRepository = doctorRepository;
             _userService = userService;
             _bookingRepository = bookingRepository;
+            _mailService = mailService;
         }
 
-        public async Task<Result<IEnumerable<User>>> GetAllDoctorsAsync(int page, int pageSize, string search, string[]? includes = null)
+        public async Task<IEnumerable<User>> GetAllDoctorsAsync(int page, int pageSize, string search, string[]? includes = null)
         {
             Expression<Func<User, bool>> searchCondition = (u) => u.Discriminator == UserDiscriminator.Doctor;
 
@@ -30,26 +39,21 @@ namespace Vezeeta.Service.Users
 
             IEnumerable<User> doctors = await _doctorRepository.FindAllAsync(searchCondition, page, pageSize, includes);
 
-            return Result.Success(doctors);
+            return doctors;
         }
 
-        public async Task<Result<User>> GetDoctorAsync(int id, string[]? includes = null)
+        public async Task<User?> GetDoctorAsync(int id, string[]? includes = null)
         {
             User? doctor = await _doctorRepository.FindAsync(d => d.Id == id && d.Discriminator == UserDiscriminator.Doctor, includes);
 
-            if (doctor == null)
-            {
-                return Result.Failure<User>(Error.Errors.Doctors.DoctorNotFound());
-            }
-
-            return Result.Success(doctor);
+            return doctor;
         }
 
-        public async Task<Result<int>> GetDoctorsCountAsync()
+        public async Task<int> GetDoctorsCountAsync()
         {
             var doctorsCount = await _doctorRepository.CountAsync(u => u.Discriminator == UserDiscriminator.Doctor);
 
-            return Result.Success(doctorsCount);
+            return doctorsCount;
         }
 
         public async Task<Result<bool>> AddDoctorAsync(User user)
@@ -59,6 +63,8 @@ namespace Vezeeta.Service.Users
                 await _userService.RegisterUserAsync(user, AppConsts.User.DoctorBasePassword, UserDiscriminator.Doctor, new[] { AppConsts.Roles.Doctor });
 
             if (registerDoctorResult.IsFailure) return Result.Failure<bool>(registerDoctorResult.Error);
+
+            await _mailService.SendAsync("Your Doctor Account Credentials", $"Email: {user.Email} Password: {AppConsts.User.DoctorBasePassword}", user);
 
             return Result.Success(true);
         }
@@ -81,6 +87,13 @@ namespace Vezeeta.Service.Users
                 return Result.Failure<bool>(Error.Errors.Doctors.DoctorHasRequests());
             }
 
+            var doctor = await GetDoctorAsync(id);
+
+            if (doctor == null)
+            {
+                return Result.Failure<bool>(Error.Errors.Doctors.DoctorNotFound());
+            }
+
             var deleteDoctorResult = await _userService.DeleteUserAsync(id);
 
             if (deleteDoctorResult.IsFailure) return Result.Failure<bool>(deleteDoctorResult.Error);
@@ -88,15 +101,31 @@ namespace Vezeeta.Service.Users
             return Result.Success(true);
         }
 
-        public async Task<Result<IEnumerable<Booking>>> GetDoctorBookings(int doctorId, int page, int pageSize, DateTime? date, string[]? includes = null)
+        public async Task<Result<bool>> UpdateDoctorPriceAsync(int doctorId, float price)
         {
-            Expression<Func<Booking, bool>> searchCondition = (b) => b.DoctorId == doctorId;
+            User? doctor = await GetDoctorAsync(doctorId);
+            if (doctor == null) return Result.Failure<bool>(Error.Errors.Doctors.DoctorNotFound());
 
-            if (date != null) searchCondition = (b) => b.DoctorId == doctorId && b.Date == date;
+            doctor.AddDoctorPrice(price);
 
-            IEnumerable<Booking> bookings = await _bookingRepository.FindAllAsync(searchCondition, page, pageSize, includes);
+            Result<bool> updateDoctorResult = await UpdateDoctorAsync(doctor);
+            if (updateDoctorResult.IsFailure) return Result.Failure<bool>(updateDoctorResult.Error);
 
-            return Result.Success(bookings);
+            return Result.Success(true);
+        }
+
+        public async Task<IEnumerable<SpecializationRequestCount>> GetTopSpecialization(int number)
+        {
+            IEnumerable<SpecializationRequestCount> specializationRequestCounts = await _bookingRepository.CountSpecializationRequestsAsync(number);
+
+            return specializationRequestCounts;
+        }
+
+        public IEnumerable<DoctorRequestCount> GetTopDoctors(int number)
+        {
+            IEnumerable<DoctorRequestCount> doctorRequestCounts = _bookingRepository.CountDoctorsRequests(number);
+
+            return doctorRequestCounts;
         }
     }
 }
